Handle sigchild, threadsafe interface
